# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fission HTTP Server API
  description: |
    REST API for managing Fission resources on Kubernetes.
    This API wraps all fission-cli commands and provides HTTP endpoints for:
    - Functions (create, list, get, update, delete, test, logs, pods, metadata)
    - Environments (create, list, get, update, delete, pods)
    - Packages (create, list, get source/deploy/info, update, delete, rebuild)
    - HTTP Triggers (create, list, get, update, delete)
    - Time Triggers (create, list, delete)
    - MQ Triggers (create, list, delete)
    - KubeWatches (create, list, delete)
    - Version information
    - System check
    - Plugins
  version: 2.0.0
  contact:
    name: Fission HTTP Server
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server (example)
paths:
...
  /api/v1/functions:
    get:
      summary: List all Fission functions
      description: |
        Retrieves a list of all Fission functions.
        Supports filtering by namespace and listing across all namespaces.
      operationId: listFunctions
      tags:
        - Functions
      parameters:
        - name: namespace
          in: query
          required: false
          description: Kubernetes namespace to filter functions
          schema:
            type: string
        - name: all_namespaces
          in: query
          required: false
          description: List functions across all namespaces
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful response with list of functions
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
                    description: Tabular output from fission-cli
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a Fission function
      description: |
        Creates a new Fission function with the specified configuration.
        Supports creating functions with code, archives, containers, and various runtime settings.
      operationId: createFunction
      tags:
        - Functions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFunctionRequest'
      responses:
        '201':
          description: Function created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  name:
                    type: string
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/functions/{name}:
    get:
      summary: Get a specific Fission function source code
      description: Retrieves the source code of a specific Fission function.
      operationId: getFunction
      tags:
        - Functions
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '404':
          description: Function not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update a Fission function
      description: Updates an existing Fission function with new configuration.
      operationId: updateFunction
      tags:
        - Functions
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFunctionRequest'
      responses:
        '200':
          description: Function updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  name:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a Fission function
      description: Deletes a Fission function by name.
      operationId: deleteFunction
      tags:
        - Functions
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          required: false
          schema:
            type: string
        - name: ignore_not_found
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Function deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  name:
                    type: string
        '404':
          description: Function not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/functions/{name}/meta:
    get:
      summary: Get function metadata
      description: Retrieves metadata information about a Fission function.
      operationId: getFunctionMeta
      tags:
        - Functions
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '404':
          description: Function not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/functions/{name}/test:
    post:
      summary: Test/invoke a Fission function
      description: Invokes a Fission function and returns the response.
      operationId: testFunction
      tags:
        - Functions
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          required: false
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: string
                headers:
                  type: object
                  additionalProperties:
                    type: string
                query:
                  type: string
                method:
                  type: string
                timeout:
                  type: integer
                subpath:
                  type: string
      responses:
        '200':
          description: Function invoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/functions/{name}/logs:
    get:
      summary: Get function logs
      description: Retrieves logs from pods running the specified Fission function.
      operationId: getFunctionLogs
      tags:
        - Functions
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          required: false
          schema:
            type: string
        - name: pod
          in: query
          required: false
          schema:
            type: string
        - name: follow
          in: query
          required: false
          schema:
            type: boolean
        - name: detail
          in: query
          required: false
          schema:
            type: string
        - name: dbtype
          in: query
          required: false
          schema:
            type: string
        - name: reverse
          in: query
          required: false
          schema:
            type: string
        - name: recordcount
          in: query
          required: false
          schema:
            type: integer
        - name: all_pods
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response with function logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/functions/{name}/pods:
    get:
      summary: List pods for a function
      description: Lists pods currently used by a Fission function.
      operationId: listFunctionPods
      tags:
        - Functions
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Environment endpoints
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error message
          example: "An error occurred"
      required:
        - error